fix(marketing): share investor deck loading
